requirements.txt
# ============================================================
# AI Answer Evaluation System — Python Dependencies
# Phase 1: Project Setup
# ============================================================

# --- Core Deep Learning ---
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# --- Computer Vision / OCR ---
opencv-python>=4.8.0
Pillow>=10.0.0
easyocr>=1.7.0
pytesseract>=0.3.10
albumentations>=1.3.0

# --- NLP / Transformers (Pretrained Model) ---
transformers>=4.35.0
sentence-transformers>=2.2.2
nltk>=3.8.1
scikit-learn>=1.3.0

# --- Data Processing ---
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# --- Backend / API ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# --- Database ---
sqlalchemy>=2.0.0
mysql-connector-python>=8.1.0
alembic>=1.12.0
pymysql>=1.1.0

# --- Utilities ---
python-dotenv>=1.0.0
tqdm>=4.66.0
PyYAML>=6.0
loguru>=0.7.0

# --- Testing ---
pytest>=7.4.0
pytest-asyncio>=0.21.0

# --- Notebook / Dev Tools ---
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.0.0
flake8>=6.1.0
